feat: #467 - Wrap ClassExpression.
