+++ src/compiler/ast/class/ClassExpression.ts
@@ -1,0 +1,7 @@
+import { ts } from "../../../typescript";
+import { ClassLikeDeclarationBase } from "./base";
+import { PrimaryExpression } from "../expression/PrimaryExpression";
+
+export const ClassExpressionBase = ClassLikeDeclarationBase(PrimaryExpression);
+export class ClassExpression extends ClassExpressionBase<ts.ClassExpression> {
+}
